+++ app/code/community/Contactlab/Template/Block/Adminhtml/Newsletter/Template/Edit/Form.php
@@ -2,10 +2,6 @@
 
 /**
  * Adminhtml Newsletter Template Edit Form Block
- *
- * @category   Mage
- * @package    Mage_Adminhtml
- * @author     Magento Core Team <core@magentocommerce.com>
  */
 class Contactlab_Template_Block_Adminhtml_Newsletter_Template_Edit_Form extends Mage_Adminhtml_Block_Newsletter_Template_Edit_Form
 {
@@ -18,19 +14,22 @@ class Contactlab_Template_Block_Adminhtml_Newsletter_Template_Edit_Form extends 
     protected function _prepareForm() {
         $returnValue = parent::_prepareForm();
 
+        /** @var $model Contactlab_Template_Model_Newsletter_Template */
         $model = $this->getModel();
         $h = Mage::helper('contactlab_template');
         $form = $this->getForm();
         $yesNo = Mage::getModel('adminhtml/system_config_source_yesno');
+        $stores = $this->getStoresOptions();
 
-        $fieldset = $form->getElement('base_fieldset');
+        /** @var $fieldSet Varien_Data_Form_Element_Fieldset */
+        $fieldSet = $form->getElement('base_fieldset');
 
         // Remove fields.
-        $fieldset->removeField('text');
-        $fieldset->removeField('template_styles');
+        $fieldSet->removeField('text');
+        $fieldSet->removeField('template_styles');
 
 
-        $fieldset->addField('enable_xml_delivery', 'select', array(
+        $fieldSet->addField('enable_xml_delivery', 'select', array(
             'required' => false,
             'name'     => 'enable_xml_delivery',
             'default'  => '1',
@@ -38,32 +37,39 @@ class Contactlab_Template_Block_Adminhtml_Newsletter_Template_Edit_Form extends 
             'values'   => $yesNo->toOptionArray(),
             'value'    => $model->getId() !== null ? $model->getEnableXmlDelivery() : 1));
 
+        $fieldSet->addField('store_id', 'select', array(
+            'required' => false,
+            'name'     => 'store_id',
+            'label'    => Mage::helper('adminhtml')->__('Store View'),
+            'values'   => $stores,
+            'value'    => $model->getStoreId()));
+
 
         $types = Mage::getModel('contactlab_template/system_config_source_template_type');
 
-        $fieldset->addField('reply_to', 'text', array(
+        $fieldSet->addField('reply_to', 'text', array(
             'required' => false,
             'name'     => 'reply_to',
             'label'    => $h->__('Reply to'),
             'value'    => $model->getId() !== null ? $model->getReplyTo() : ''));
-        $fieldset->addField('template_type_id', 'select', array(
+        $fieldSet->addField('template_type_id', 'select', array(
             'required' => false,
             'name'     => 'template_type_id',
             'values'   => $types->toOptionArray(),
             'label'    => $h->__('Template type'),
             'value'    => $model->getId() !== null ? $model->getTemplateTypeId() : ''));
-        $fieldset->addField('text', 'textarea', array(
+        $fieldSet->addField('text', 'textarea', array(
             'required' => true,
             'name'     => 'text',
             'label'    => $h->__('Template (html format)'),
             'value'    => $model->getTemplateText()));
-        $fieldset->addField('template_text_plain', 'textarea', array(
+        $fieldSet->addField('template_text_plain', 'textarea', array(
             'required' => true,
             'name'     => 'template_text_plain',
             'label'    => $h->__('Template (text format)'),
             'value'    => $model->getTemplateTextPlain()));
         $formats = Mage::getModel('contactlab_template/system_config_source_template_format');
-        $fieldset->addField('flg_html_txt', 'select', array(
+        $fieldSet->addField('flg_html_txt', 'select', array(
             'required' => true,
             'name'     => 'flg_html_txt',
             'label'    => $h->__('Email format type'),
@@ -71,13 +77,13 @@ class Contactlab_Template_Block_Adminhtml_Newsletter_Template_Edit_Form extends 
             'values'   => $formats->toOptionArray(),
             'value'    => $model->getId() !== null ? $model->getFlgHtmlTxt() : 'B'));
 
-        $fieldset->addField('template_styles', 'textarea', array(
+        $fieldSet->addField('template_styles', 'textarea', array(
             'name'          =>'styles',
             'label'         => Mage::helper('newsletter')->__('Template Styles'),
             'container_id'  => 'field_template_styles',
             'value'         => $model->getTemplateStyles()
         ));
-        $fieldset->addField('is_test_mode', 'select', array(
+        $fieldSet->addField('is_test_mode', 'select', array(
             'name'      => 'is_test_mode',
             'label'     => Mage::helper('contactlab_template')->__('Send to test recipients'),
             'title'     => Mage::helper('contactlab_template')->__('Send to test recipients'),
@@ -89,9 +95,9 @@ class Contactlab_Template_Block_Adminhtml_Newsletter_Template_Edit_Form extends 
         ));
 
 
-        $this->_addProductFieldset();
+        $this->_addProductFieldSet();
         $this->_addCustomerFilterOptions();
-        $this->_addCronFieldset();
+        $this->_addCronFieldSet();
 
         return $returnValue;
     }
@@ -100,24 +106,25 @@ class Contactlab_Template_Block_Adminhtml_Newsletter_Template_Edit_Form extends 
     /**
      * Add product fields fileset
      */
-    private function _addProductFieldset() {
+    private function _addProductFieldSet() {
+        /** @var $model Contactlab_Template_Model_Newsletter_Template */
         $model = $this->getModel();
         $h = Mage::helper('contactlab_template');
         $form = $this->getForm();
 
-        $fieldset = $form->addFieldset('product_fieldset', array(
+        $fieldSet = $form->addFieldset('product_fieldset', array(
             'legend' => $h->__('Product templates'),
             'class'  => 'fieldset-wide'
         ));
 
         foreach (range(1, 5) as $i) {
-            $fieldset->addField("template_pr_txt_$i", 'textarea', array(
+            $fieldSet->addField("template_pr_txt_$i", 'textarea', array(
                 'required' => false,
                 'name'     => "template_pr_txt_$i",
                 'label'    => $h->__("Product template nr %d (text)", $i),
                 'title'    => $h->__("Template nr %d for product (text/plain)", $i),
                 'value'    => $model->getId() !== null ? $model->getData("template_pr_txt_$i") : ''));
-            $fieldset->addField("template_pr_html_$i", 'textarea', array(
+            $fieldSet->addField("template_pr_html_$i", 'textarea', array(
                 'required' => false,
                 'name'     => "template_pr_html_$i",
                 'label'    => $h->__("Product template nr %d (html)", $i),
@@ -125,14 +132,14 @@ class Contactlab_Template_Block_Adminhtml_Newsletter_Template_Edit_Form extends 
                 'value'    => $model->getId() !== null ? $model->getData("template_pr_html_$i") : ''));
         }
 
-        $fieldset->addField('default_product_snippet', 'text', array(
+        $fieldSet->addField('default_product_snippet', 'text', array(
             'required' => false,
             'name'     => 'default_product_snippet',
             'label'    => $h->__('Default product snippet (1-5)'),
             'title'    => $h->__('Default product snippet number'),
             'value'    => $model->getId() !== null ? $model->getDefaultProductSnippet() : ''));
 
-        $fieldset->addField('product_image_size', 'text', array(
+        $fieldSet->addField('product_image_size', 'text', array(
             'required' => false,
             'name'     => 'product_image_size',
             'label'    => $h->__('Product image size (width x height)'),
@@ -144,26 +151,26 @@ class Contactlab_Template_Block_Adminhtml_Newsletter_Template_Edit_Form extends 
     /**
      * Add cron fields fileset
      */
-    private function _addCronFieldset() {
+    private function _addCronFieldSet() {
+        /** @var $model Contactlab_Template_Model_Newsletter_Template */
         $model = $this->getModel();
         $h = Mage::helper('contactlab_template');
         $form = $this->getForm();
         $yesNo = Mage::getModel('adminhtml/system_config_source_yesno');
 
-        $form = $this->getForm();
-        $fieldset = $form->addFieldset('cron_fieldset', array(
+        $fieldSet = $form->addFieldset('cron_fieldset', array(
             'legend' => $h->__('Cron Information'),
             'class'  => 'fieldset-narrow'
         ));
 
-        $fieldset->addField('is_cron_enabled', 'select', array(
+        $fieldSet->addField('is_cron_enabled', 'select', array(
             'label'      => $h->__('Activate for Cron execution'),
             'title'      => $h->__('Does the template is active for Cron?'),
             'name'       => 'is_cron_enabled',
             'values'   => $yesNo->toOptionArray(),
             'value'    => $model->getId() !== null ? $model->getIsCronEnabled() : 0));
         $dateFormatIso = Mage::app()->getLocale()->getDateFormat(Mage_Core_Model_Locale::FORMAT_TYPE_SHORT);
-        $fieldset->addField('cron_date_range_start', 'date', array(
+        $fieldSet->addField('cron_date_range_start', 'date', array(
             'label'        => $h->__('Enabled start date'),
             'title'        => $h->__('Start date for Cron'),
             'image'        => $this->getSkinUrl('images/grid-cal.gif'),
@@ -172,7 +179,7 @@ class Contactlab_Template_Block_Adminhtml_Newsletter_Template_Edit_Form extends 
             'required'     => false,
             'name'         => 'cron_date_range_start',
             'value'        => $model->getId() !== null ? $model->getCronDateRangeStart() : ''));
-        $fieldset->addField('cron_date_range_end', 'date', array(
+        $fieldSet->addField('cron_date_range_end', 'date', array(
             'label'        => $h->__('Enabled end date'),
             'title'        => $h->__('End date for Cron'),
             'image'        => $this->getSkinUrl('images/grid-cal.gif'),
@@ -181,14 +188,14 @@ class Contactlab_Template_Block_Adminhtml_Newsletter_Template_Edit_Form extends 
             'required'     => false,
             'name'         => 'cron_date_range_end',
             'value'        => $model->getId() !== null ? $model->getCronDateRangeEnd() : ''));
-        $fieldset->addField('queue_delay_time', 'text', array(
+        $fieldSet->addField('queue_delay_time', 'text', array(
             'label'    => $h->__('Queue delay time'),
             'title'    => $h->__('Optional queue delay time'),
             'required' => false,
             'name'     => 'queue_delay_time',
             'value'    => $model->getId() !== null ? $model->getQueueDelayTime() : 0));
 
-        $fieldset->addField('priority', 'text', array(
+        $fieldSet->addField('priority', 'text', array(
             'required' => true,
             'name'     => 'priority',
             'label'    => $h->__('Priority'),
@@ -201,18 +208,20 @@ class Contactlab_Template_Block_Adminhtml_Newsletter_Template_Edit_Form extends 
      * @return $this
      */
     public function _addCustomerFilterOptions() {
+        /** @var $model Contactlab_Template_Model_Newsletter_Template */
         $model = $this->getModel();
+
         /* @var $h Contactlab_Template_Helper_Data */
         $h = Mage::helper('contactlab_template');
         $form = $this->getForm();
         $andOr = Mage::getModel('contactlab_template/system_config_source_andOr');
 
-        $fieldset = $form->addFieldset('customer_filter_fieldset', array(
+        $fieldSet = $form->addFieldset('customer_filter_fieldset', array(
             'legend' => $h->__('Customer filter options'),
             'class'  => 'fieldset-narrow'
         ));
 
-        $fieldset->addField('min_minutes_from_last_update', 'text', array(
+        $fieldSet->addField('min_minutes_from_last_update', 'text', array(
             'required' => false,
             'class' => 'validate-not-negative-number validate-digits validate-is-min-of-range',
             'label' => $h->__('Minimum number of minutes'),
@@ -220,7 +229,7 @@ class Contactlab_Template_Block_Adminhtml_Newsletter_Template_Edit_Form extends 
             'name' => 'min_minutes_from_last_update',
             'value'    => $model->getId() !== null ? $model->getMinMinutesFromLastUpdate() : ''));
 
-        $fieldset->addField('max_minutes_from_last_update', 'text', array(
+        $fieldSet->addField('max_minutes_from_last_update', 'text', array(
             'required' => false,
             'class' => 'validate-not-negative-number validate-digits validate-is-max-of-range',
             'label' => $h->__('Maximum number of minutes'),
@@ -228,7 +237,7 @@ class Contactlab_Template_Block_Adminhtml_Newsletter_Template_Edit_Form extends 
             'name' => 'max_minutes_from_last_update',
             'value'    => $model->getId() !== null ? $model->getMaxMinutesFromLastUpdate() : ''));
 
-        $fieldset->addField('min_value', 'text', array(
+        $fieldSet->addField('min_value', 'text', array(
             'required' => false,
             'class' => 'validate-number validate-not-negative-number validate-is-min-of-range',
             'label' => $h->__('Minimum value'),
@@ -237,7 +246,7 @@ class Contactlab_Template_Block_Adminhtml_Newsletter_Template_Edit_Form extends 
             'name' => 'min_value',
             'value'    => $model->getId() !== null ? $h->formatPrice($model->getMinValue()) : ''));
 
-        $fieldset->addField('max_value', 'text', array(
+        $fieldSet->addField('max_value', 'text', array(
             'required' => false,
             'class' => 'validate-number validate-not-negative-number validate-is-max-of-range',
             'label' => $h->__('Maximum value'),
@@ -245,7 +254,7 @@ class Contactlab_Template_Block_Adminhtml_Newsletter_Template_Edit_Form extends 
             'name' => 'max_value',
             'value'    => $model->getId() !== null ? $h->formatPrice($model->getMaxValue()) : ''));
 
-        $fieldset->addField('min_products', 'text', array(
+        $fieldSet->addField('min_products', 'text', array(
             'required' => false,
             'class' => 'validate-greater-than-zero validate-digits validate-is-min-of-range',
             'label' => $h->__('Minimum number of products'),
@@ -253,7 +262,7 @@ class Contactlab_Template_Block_Adminhtml_Newsletter_Template_Edit_Form extends 
             'name' => 'min_products',
             'value'    => $model->getId() !== null ? $model->getMinProducts() : ''));
 
-        $fieldset->addField('max_products', 'text', array(
+        $fieldSet->addField('max_products', 'text', array(
             'required' => false,
             'class' => 'validate-greater-than-zero validate-digits validate-is-max-of-range',
             'label' => $h->__('Maximum number of products'),
@@ -261,7 +270,7 @@ class Contactlab_Template_Block_Adminhtml_Newsletter_Template_Edit_Form extends 
             'name' => 'max_products',
             'value'    => $model->getId() !== null ? $model->getMaxProducts() : ''));
 
-        $fieldset->addField('and_or', 'select', array(
+        $fieldSet->addField('and_or', 'select', array(
             'required' => true,
             'label' => $h->__('And/or condition values'),
             'title' => $h->__('And/or condition values'),
@@ -270,5 +279,20 @@ class Contactlab_Template_Block_Adminhtml_Newsletter_Template_Edit_Form extends 
             'value'    => $model->getId() !== null ? $model->getAndOr() : 'AND'));
 
         return $this;
+    }
+
+    /**
+     * Get stores options.
+     * @return array
+     */
+    private function getStoresOptions()
+    {
+        /* @var $h Contactlab_Template_Helper_Data */
+        $h = Mage::helper('contactlab_template');
+        $stores = Mage::getModel('adminhtml/system_config_source_store')->toOptionArray();
+        $stores = array_reverse($stores);
+        $stores['none'] = $h->__('Any');
+        $stores = array_reverse($stores);
+        return $stores;
     }
 }