+++ CHANGELOG.md
@@ -1,9 +1,14 @@
 # Changelog 
 
 
+## 2.2.6 (2015-09-29)
+
+- Templates may now be associated to a specific store view ([831d375](https://github.com/contactlab/contactlab-magento-connect/commit/831d375c550d4f52a2c361d71d55a1aee5193443))
+
+
 ## 2.2.5 (2015-09-08)
 
-- ContactLab Connect for Magento is now open source on [https://github.com/contactlab/contactlab-magento-connec](https://github.com/contactlab/contactlab-magento-connect).
+- ContactLab Connect for Magento is now open source on [https://github.com/contactlab/contactlab-magento-connect](https://github.com/contactlab/contactlab-magento-connect).
 - Per store subscription status is supported in multi store configurations.
 - Increased export capabilities for large data sets.
 - Templates size limit constraint removed.